requirements.txt
langchain
openai
tiktoken
faiss-cpu
langchain_experimental
langchain[docarray]